// dns/models/DnsRecordResponse.ts
/**
 * DNS
 * Manage dns zones and records
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@liara.ir
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DnsRecordId } from '../models/DnsRecordId';
import { HttpFile } from '../http/http';

export class DnsRecordResponse {
    /**
    * success
    */
    'status'?: string;
    'data'?: DnsRecordId;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "string",
            "format": ""
        },
        {
            "name": "data",
            "baseName": "data",
            "type": "DnsRecordId",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DnsRecordResponse.attributeTypeMap;
    }

    public constructor() {
    }
}